Removed the conditional operator from the JSX
Expeses.js: The conditional operator was removed from JSX and an if
statement was added in the top. the if state checks the length of the
array and then dispaly the content accordingly. the content displayed
adjusts according to the date selected and a message is diplayed if
there are no matches.